+++ Client/Controllers/ClientHub.cs
@@ -65,7 +65,7 @@ namespace FreeTabletop.Client.Controllers
                 Networker.hubConnection.On<int, string>("Tabletop:SyncCells", Room.SyncCells);
                 Networker.hubConnection.On<string, int[]>("Tabletop:UpdateEntityPosition", UpdateEntityPosition);
                 Networker.hubConnection.On<bool>("Tabletop:UpdateLock", Room.UpdateLock);
-                Networker.hubConnection.On<string>("Tabletop:LoadPopupImage", RenderPopupImage);
+                Networker.hubConnection.On<string, string>("Tabletop:LoadPopupImage", RenderPopupImage);
                 Networker.hubConnection.On<bool>("Tabletop:ToggleVisibility", Room.SetTabletopVisibility);
 
                 Networker.hubConnection.On<string>("Notification:PlayerConnected", ConnectedNotification);
@@ -314,14 +314,14 @@ namespace FreeTabletop.Client.Controllers
             JSRuntime.InvokeVoidAsync("UpdateEntityPosition", uid, position, Tabletop.CellSize);
         }
 
-        public void LoadPopupImage(string url)
+        public void LoadPopupImage(string url, string label)
         {
-            Networker.hubConnection.SendAsync("Room:LoadPopupImage", url);
+            Networker.hubConnection.SendAsync("Room:LoadPopupImage", url, label);
         }
 
-        public void RenderPopupImage(string url)
+        public void RenderPopupImage(string url, string label)
         {
-            JSRuntime.InvokeVoidAsync("RenderPopupImage", url);
+            JSRuntime.InvokeVoidAsync("RenderPopupImage", url, label);
         }
 
         public void SetBleeding(string uid, bool isBleeding)